Update answer view for document upload
Support uploading documents for a 'file' question and general
document uploading. If there is a file upload question, it displays
any files associated with it and has an upload box. In addition,
there is a general file upload section at the bottom.

This fix also changes the upload interface from ajax autsave to a
submit button. The ajax saving is intuitive for the question text
fields, but not the file uploads. The page didn't update with the
file uploaded and it was easy to upload a file multiple times by
changing focus on the upload box.

Signed-off-by: Marc Jones <marcj303@gmail.com>