Add file upload to the application review page
Allow users and judges to upload files to the application
after it has been submitted. This allows files to be uploaded
in the Q&A period and could could allow judges to upload contract
docs, etc later in the process. Note, the users that uploads
a file is the only one that can delete it.

Signed-off-by: Marc Jones <marcj303@gmail.com>