Default feedback type to textarea
Judges were using the default test input area, which was often too
small an area for the applicant to answer. The text input isn't limited
in length, but the text box is and the answer is unreadable if the
user types something longer than the box displayed for the judge.

Signed-off-by: Marc Jones <marcj303@gmail.com>
(cherry picked from commit 72d7d5887adfa90f68d54f8a9815a6310ea69632)
Signed-off-by: Marc Jones <marcj303@gmail.com>